Postgres supports nested transactions by way of savepoints.
http://www.postgresql.org/docs/8.1/static/sql-savepoint.html

An example usage is

BEGIN;
    INSERT INTO table1 VALUES (1);
    SAVEPOINT my_savepoint;
    INSERT INTO table1 VALUES (2);
    ROLLBACK TO SAVEPOINT my_savepoint;
    INSERT INTO table1 VALUES (3);
COMMIT;

This patch adds savepoint support to the peewee PostgresqlDatabase. As is, PostgresqlDatabase doesn't implement begin, or support nested transactions.

begin: now creates a savepoint.
commit: releases the savepoint, and calls commit if not still nested.
rollback: rolls back to the savepoint, and calls rollback if not still nested.