Added support for introspecting views for Postgresql.
A command line switch will introspect views instead of tables in pwiz.py. A
get_views function was added for Postgres and NotImplemented version for the
interface. It would probably be more convenient for users to do both tables and
views at once, but this was an extremely simple way to do it. I did not add it for
others since I am mostly familiar with Postgres, but it seems pretty easy.