Implement text-based decimal field for Sqlite.
Sqlite uses double-precision floats to store decimal values, which can
result in truncation or other quirks when using particular values. This
patch adds a playhouse.sqlite_ext.TDecimalField subclass which stores
the value in a text-affinity column, which preserves the original value.
The caveat is that attempting to do math in the database will result in
conversion/possible truncation.

Refs #2008