Fix for issue #1991 regarding setting intervening models to None.
This particular issue involves joining 3 models together in a chain,
where the outer two models are empty. Previously peewee would make the
middle model an empty model instance (since a link might be needed from
the source model to the outermost model). But since both were empty, it
is more correct to make the intervening model a NULL value on the
foreign-key field rather than an empty instance.