Delegate to the ForeignKeyField's db_value() method when converting models.
When a model instance is passed as a parameter on the right-hand-side of
an expression, Peewee would previously just grab the model's primary-key
value. This breaks queries where the left-hand-side is a foreign-key
that points to a field other than the primary-key.

To address this I've added a test to try and use the converter in the
scope (if available), and fall back to using the model pk.

Fixes #2304