Attempt new connections if the server is not ready
When using `ReconnectMixin`, the execution of a SQL statement is attempted again if it fails because of "operational issue" (for example when the connection to the DB server is lost). During the reconnection phase, the DB server might not be ready to handle new connections, thus the call to `self.connect()` could also fail.