Reverts ab43376 / #2026 in favor of #2556 for bulk inserts
This change is backwards incompatible and affects the default value
returned by bulk insert operations on MySQL and Sqlite (when no
RETURNING clause is specified).

Prior to this commit, Peewee has (since 3.12) returned the modified
row-count rather than the cursor lastrowid. This change returns Peewee
to its previous behavior of returning the lastrowid, and exposes the
modified rowcount using a new `as_rowcount()` chainable method on the
Insert query object.