Comment-out invalid as_rowcount() + RETURNING tests for sqlite.
The `sqlite3_changes()` interface does not populate the changes until
the statement containing the RETURNING has been consumed. The only
reason these tests were passing in the first place is likely because
they happened to be consumed in the first assertion, and then sat in the
statement cache and were re-run in the second assertion that checked the
rowcount.