Run psycopg2 in autocommit and fully take over txn management.
This is similar to what we currently do for sqlite3 - rather than
relying on the db-api transaction management under-the-hood, calling
commit as necessary, we take over the transaction management ourselves.
The semantics should be the same, but ideally with less performance
impact as the driver is issuing fewer BEGIN/COMMITs. This also has the
benefit of not leaving connections in the "InternalError: Current
transaction is aborted" state for "top-level" queries outside of
transactions.

This also removes the db-api-specific cruft in execute_sql() and assumes
that overrides will be put into autocommit-mode. We'll retain the db-api
stuff for MySQL for now, at any rate.