Fix threadsafety issue when using various decorators.
When the decorator was used only a single instance was created. Rather
than storing state in a threadlocal, we'll have the decorators instead
create a new instance to use a context manager internally.

This hurts some of the reuse since the constructors for these different
classes take different arguments, but the end result is working well.

Fixes #2709