Do not special-case JSONB contains for scalar keys.
Previously, passing a string/int/bool to BinaryJSONField.contains()
would silently switch to using the EXISTS (?) operator to test for a
key. This is somewhat bad behavior (#2984) and so the special-case is
removed.

For the old behavior of testing for presence of a key/item at the
top-level, you should use the .has_key() method instead.